wip for multi user support

+++ lib/qbwc/qbwc_session.rb
@@ -10,7 +10,7 @@ module QBWC
     def current_request
       request = nil
       if job = self.next_job ||= next_job_in_queue
-        obj = eval(job.klass).send(:find, job.klass_id)
+        obj = job.klass.constantize.send(:find, job.klass_id)
         request = obj.qb_payload
         request = QBWC::Request.new(request)
         advance
@@ -19,7 +19,8 @@ module QBWC
     end
 
     def next_job_in_queue
-      jobs = QBWC::QbwcJob.where(processed: false).order('id asc')
+      jobs = QBWC::QbwcJob.where(processed: false, owner_id: self.owner_id, owner_type: self.owner_type).order('id asc')
+      jobs = jobs.where(owner_id: self.owner_id, owner_type: self.owner_type)
       jobs = jobs.where('id != ?', self.next_job.id) if self.next_job.present?
       jobs.limit(1).first
     end
@@ -31,20 +32,20 @@ module QBWC
     end
 
     def progress
-      n_processed = QBWC::QbwcJob.where(processed: true).count.to_f
+      n_processed = QBWC::QbwcJob.where(processed: true, owner_id: self.owner_id, owner_type: self.owner_type).count.to_f
       total_jobs = QBWC::QbwcJob.count.to_f
       return 100 if n_processed == total_jobs
       ((n_processed / total_jobs)*100).to_i
     end
 
     def complete_session
-      QBWC::QbwcJob.where(processed: true).destroy_all
+      QBWC::QbwcJob.where(processed: true, owner_id: self.owner_id, owner_type: self.owner_type).destroy_all
       self.destroy
     end
 
     private
     def setup
-      self.ticket = Digest::SHA1.hexdigest("#{Rails.application.config.secret_token}#{Time.now.to_i}")
+      self.ticket = SecureRandom.uuid#Digest::SHA1.hexdigest("#{Rails.application.config.secret_token}#{Time.now.to_i}")
     end
   end
 end
